requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
absl-py=1.0.0=pyhd8ed1ab_0
blas=2.113=mkl
blas-devel=3.9.0=13_linux64_mkl
bzip2=1.0.8=h7f98852_4
c-ares=1.18.1=h7f98852_0
ca-certificates=2021.10.8=ha878542_0
cairo=1.14.12=h80bd089_1005
cloudpickle=2.0.0=pyhd8ed1ab_0
cudatoolkit=11.0.3=h15472ef_10
cycler=0.11.0=pyhd8ed1ab_0
cytoolz=0.11.0=py36h8f6f2f9_3
dask-core=2021.3.0=pyhd8ed1ab_0
dataclasses=0.8=pyh787bdff_2
decorator=5.1.1=pyhd8ed1ab_0
ffmpeg=3.4.1=0
fontconfig=2.13.1=he4413a7_1000
freetype=2.10.4=h0708190_1
gettext=0.19.8.1=hf34092f_1004
glib=2.56.2=had28632_1001
graphite2=1.3.13=h58526e2_1001
grpcio=1.16.0=py36h4f00d22_1000
harfbuzz=1.9.0=he243708_1001
hdf5=1.10.1=2
icu=58.2=hf484d3e_1000
imagecodecs-lite=2019.12.3=py36h92226af_3
imageio=2.13.1=pyhd8ed1ab_0
jasper=1.900.1=4
jpeg=9b=0
kiwisolver=1.3.1=py36h605e78d_1
libblas=3.9.0=13_linux64_mkl
libcblas=3.9.0=13_linux64_mkl
libffi=3.2.1=he1b5a44_1007
libgcc-ng=11.2.0=h1d223b6_12
libgfortran=3.0.0=1
libgfortran-ng=11.2.0=h69a702a_12
libgfortran5=11.2.0=h5c6108e_12
libiconv=1.15=h516909a_1006
liblapack=3.9.0=13_linux64_mkl
liblapacke=3.9.0=13_linux64_mkl
libpng=1.6.37=h21135ba_2
libprotobuf=3.18.0=h780b84a_1
libstdcxx-ng=11.2.0=he4da1e4_12
libtiff=4.0.9=he6b73bb_1
libuuid=2.32.1=h7f98852_1000
libuv=1.43.0=h7f98852_0
libxcb=1.13=h7f98852_1004
libxml2=2.9.9=h13577e0_2
libzlib=1.2.11=h36c2ea0_1013
llvm-openmp=13.0.1=hf817b99_0
markdown=3.2.1=py_0
matplotlib-base=3.1.1=py36hfd891ef_0
mkl=2022.0.1=h8d4b97c_803
mkl-devel=2022.0.1=ha770c72_804
mkl-include=2022.0.1=h8d4b97c_803
ncurses=6.3=h9c3ff4c_0
networkx=2.5=py_0
ninja=1.10.2=h4bd325d_1
numpy=1.19.5=py36hfc0c790_2
olefile=0.46=pyh9f0ad1d_1
opencv=3.3.1=py36h0a11808_0
openssl=1.0.2u=h516909a_0
pcre=8.45=h9c3ff4c_0
pillow=5.4.1=py36h34e0f95_0
pip=21.3.1=pyhd8ed1ab_0
pixman=0.34.0=h14c3975_1003
protobuf=3.18.0=py36hc4f0c31_0
pthread-stubs=0.4=h36c2ea0_1001
pyparsing=3.0.7=pyhd8ed1ab_0
python=3.6.6=hd21baee_1003
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.6=2_cp36m
pytorch=1.7.1=py3.6_cuda11.0.221_cudnn8.0.5_0
pywavelets=1.1.1=py36he33b4a0_3
pyyaml=5.4.1=py36h8f6f2f9_1
readline=7.0=hf8c457e_1001
scikit-image=0.17.2=py36h284efc9_4
scipy=1.5.3=py36h81d768a_1
setuptools=58.0.4=py36h5fab9bb_2
six=1.16.0=pyh6c4a22f_0
sqlite=3.28.0=h8b20d00_0
tbb=2021.5.0=h4bd325d_0
tensorboard=1.15.0=py36_0
tensorboardx=2.5=pyhd8ed1ab_0
tifffile=2020.6.3=py_0
tk=8.6.12=h27826a3_0
toolz=0.11.2=pyhd8ed1ab_0
torchaudio=0.7.2=py36
torchvision=0.8.2=py36_cu110
tornado=6.1=py36h8f6f2f9_1
typing_extensions=4.1.1=pyha770c72_0
werkzeug=2.0.2=pyhd8ed1ab_0
wheel=0.37.1=pyhd8ed1ab_0
x264=20131218=0
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.3=hd9c2040_1000
xorg-libx11=1.7.2=h7f98852_0
xorg-libxau=1.0.9=h7f98852_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h516909a_1
yaml=0.2.5=h7f98852_2
zlib=1.2.11=h36c2ea0_1013
